+++ lib/vue-form-for.rb
@@ -1,2 +1,0 @@
-require "vue-form-for/form_builder"
-require "vue-form-for/railtie"