+++ lib/vue-rails-form-builder.rb
@@ -1,0 +1,2 @@
+require "vue-rails-form-builder/form_builder"
+require "vue-rails-form-builder/railtie"