+++ lib/vue-rails-form-builder/form_helper.rb
@@ -1,7 +1,7 @@
-module VueFormFor
+module VueRailsFormBuilder
   module FormHelper
     def vue_form_for(record, options = {}, &block)
-      options[:builder] ||= VueFormFor::FormBuilder
+      options[:builder] ||= VueRailsFormBuilder::FormBuilder
       form_for(record, options, &block)
     end
   end