+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# CHANGELOG - vue-form-for
+# CHANGELOG - vue-rails-form-builder
+
+## 0.5.0 (2017-05-12)
+
+* Change the name of this gem to `vue-rails-form-builder` from `vue-form-for`
 
 ## 0.4.0 (2017-05-11)